Python 3 + Mixbox Support

+++ cybox/test/objects/memory_test.py
@@ -3,7 +3,11 @@
 
 import unittest
 
+from mixbox.vendor.six import u
+
 from cybox.objects.memory_object import Memory
+
+from cybox.compat import long
 from cybox.test.common.hash_test import EMPTY_MD5
 from cybox.test.objects import ObjectTestCase
 
@@ -21,14 +25,14 @@ class TestMemory(ObjectTestCase, unittest.TestCase):
         'is_mapped': False,
         'is_protected': True,
         'is_volatile': False,
-        'hashes': [{'type': u"MD5", 'simple_hash_value': EMPTY_MD5}],
-        'name': u"A memory region",
-        'region_size': 65536L,
-        'memory_source': u".data",
-        'block_type': u"Free",
-        'region_start_address': u"00040000",
-        'region_end_address': u"00048000",
-        'extracted_features': {'functions': [u"StringA", u"StringB"]},
+        'hashes': [{'type': u("MD5"), 'simple_hash_value': EMPTY_MD5}],
+        'name': u("A memory region"),
+        'region_size': long(65536),
+        'memory_source': u(".data"),
+        'block_type': u("Free"),
+        'region_start_address': u("00040000"),
+        'region_end_address': u("00048000"),
+        'extracted_features': {'functions': [u("StringA"), u("StringB")]},
         'xsi:type': object_type,
     }
