+++ cybox/test/objects/win_memory_page_region_test.py
@@ -3,8 +3,11 @@
 
 import unittest
 
+from mixbox.vendor.six import u
+
 from cybox.objects.win_memory_page_region_object import WinMemoryPageRegion
 
+from cybox.compat import long
 from cybox.test import EntityTestCase, round_trip
 from cybox.test.objects import ObjectTestCase
 
@@ -18,17 +21,17 @@ class TestWinMemoryPageRegion(ObjectTestCase, unittest.TestCase):
         'is_mapped': False,
         'is_protected': True,
         'is_volatile': False,
-        'name': u"A page region",
-        'memory_source': u"A source",
-        'region_size': 10000L,
-        'block_type': u"A block type",
-        'region_start_address': u"1234abcde",
-        'region_end_address': u"1234abdde",
-        'allocation_base_address': u"1234abbbe",
-        'type': u"A type",
-        'allocation_protect': u"allocate protection",
-        'state': u"A state",
-        'protect': u"protection",
+        'name': u("A page region"),
+        'memory_source': u("A source"),
+        'region_size': long(10000),
+        'block_type': u("A block type"),
+        'region_start_address': u("1234abcde"),
+        'region_end_address': u("1234abdde"),
+        'allocation_base_address': u("1234abbbe"),
+        'type': u("A type"),
+        'allocation_protect': u("allocate protection"),
+        'state': u("A state"),
+        'protect': u("protection"),
         #'extracted_features' = cybox.TypedField("Extracted_Features", ExtractedFeatures)
         #'hashes' = cybox.TypedField("Hashes", HashList)
         'xsi:type': object_type,
